Start on artifact tests.
