+++ package.json
@@ -41,7 +41,6 @@
     "@types/node": "^16.11.58",
     "@types/npm-packlist": "^3.0.0",
     "@types/react": "^17.0.48",
-    "@types/rimraf": "^3.0.2",
     "@types/semver": "^7.3.12",
     "babel-preset-moon": "^1.0.1",
     "chokidar": "^3.5.3",
