+++ packages/packemon/package.json
@@ -99,7 +99,6 @@
     "npm-packlist": "^5.1.1",
     "react": "^17.0.2",
     "resolve": "^1.22.1",
-    "rimraf": "^3.0.2",
     "rollup": "^2.79.0",
     "rollup-plugin-node-externals": "^4.1.1",
     "rollup-plugin-polyfill-node": "^0.10.2",
