+++ packages/packemon/tests-old/CodeArtifact.test.ts
@@ -35,16 +35,7 @@ describe('CodeArtifact', () => {
 			new Package(new Project(fixturePath), fixturePath, { ...packageJson }),
 			[],
 		);
-		artifact.inputs = { index: 'src/index.ts' };
 		artifact.startup();
-	});
-
-	it('sets correct metadata', () => {
-		artifact.builds.push({ format: 'lib' }, { format: 'cjs' });
-
-		expect(artifact.getLabel()).toBe('node:stable:lib,cjs');
-		expect(artifact.getBuildTargets()).toEqual(['lib', 'cjs']);
-		expect(artifact.toString()).toBe('code (node:stable:lib,cjs)');
 	});
 
 	describe('build()', () => {
@@ -101,109 +92,6 @@ describe('CodeArtifact', () => {
 
 			expect(bundleWriteSpy).toHaveBeenCalledWith({ c: true });
 			expect(artifact.builds[2].stats?.size).toBe(4);
-		});
-	});
-
-	describe('getInputPaths()', () => {
-		it('returns an absolute path for every input', () => {
-			expect(artifact.getInputPaths()).toEqual({
-				index: fixturePath.append('src/index.ts').path(),
-			});
-		});
-	});
-
-	describe('getBuildOutput()', () => {
-		beforeEach(() => {
-			artifact.platform = 'node';
-		});
-
-		it('returns metadata for `lib` format', () => {
-			expect(artifact.getBuildOutput('lib', 'index')).toEqual({
-				ext: 'js',
-				extGroup: 'js,map',
-				file: 'index.js',
-				folder: 'lib',
-				path: './lib/index.js',
-			});
-		});
-
-		it('returns metadata for `esm` format', () => {
-			expect(artifact.getBuildOutput('esm', 'index')).toEqual({
-				ext: 'js',
-				extGroup: 'js,map',
-				file: 'index.js',
-				folder: 'esm',
-				path: './esm/index.js',
-			});
-		});
-
-		it('returns metadata for `umd` format', () => {
-			expect(artifact.getBuildOutput('umd', 'index')).toEqual({
-				ext: 'js',
-				extGroup: 'js,map',
-				file: 'index.js',
-				folder: 'umd',
-				path: './umd/index.js',
-			});
-		});
-
-		it('returns metadata for `cjs` format', () => {
-			expect(artifact.getBuildOutput('cjs', 'index')).toEqual({
-				ext: 'cjs',
-				extGroup: 'cjs,mjs,map',
-				file: 'index.cjs',
-				folder: 'cjs',
-				path: './cjs/index.cjs',
-			});
-		});
-
-		it('returns metadata for `mjs` format', () => {
-			expect(artifact.getBuildOutput('mjs', 'index')).toEqual({
-				ext: 'mjs',
-				extGroup: 'mjs,map',
-				file: 'index.mjs',
-				folder: 'mjs',
-				path: './mjs/index.mjs',
-			});
-		});
-
-		describe('shared lib', () => {
-			it('includes platform in folder when shared lib required', () => {
-				artifact.sharedLib = true;
-
-				expect(artifact.getBuildOutput('lib', 'index')).toEqual({
-					ext: 'js',
-					extGroup: 'js,map',
-					file: 'index.js',
-					folder: 'lib/node',
-					path: './lib/node/index.js',
-				});
-			});
-
-			it('ignores shared lib if not `lib` format', () => {
-				artifact.sharedLib = true;
-
-				expect(artifact.getBuildOutput('esm', 'index')).toEqual({
-					ext: 'js',
-					extGroup: 'js,map',
-					file: 'index.js',
-					folder: 'esm',
-					path: './esm/index.js',
-				});
-			});
-		});
-
-		it('public api: uses original source file path', () => {
-			artifact.api = 'public';
-			artifact.inputs = { index: 'src/some/other/file.ts' };
-
-			expect(artifact.getBuildOutput('mjs', 'index')).toEqual({
-				ext: 'mjs',
-				extGroup: 'mjs,map',
-				file: 'some/other/file.mjs',
-				folder: 'mjs',
-				path: './mjs/some/other/file.mjs',
-			});
 		});
 	});
