+++ packages/packemon/tests-old/TypesArtifact.test.ts
@@ -46,117 +46,6 @@ describe('TypesArtifact', () => {
 		expect(artifact.toString()).toBe('types (dts)');
 	});
 
-	describe('build()', () => {
-		let declSpy: jest.SpyInstance;
-		let declRootSpy: jest.SpyInstance;
-
-		beforeEach(() => {
-			declSpy = jest.spyOn(artifact, 'generateDeclarations');
-			declRootSpy = jest.spyOn(artifact.package.project, 'generateDeclarations');
-		});
-
-		afterEach(() => {
-			declSpy.mockRestore();
-			declRootSpy.mockRestore();
-		});
-
-		it('without refs, runs project `tsc` without --build', async () => {
-			await artifact.build({});
-
-			expect(declRootSpy).toHaveBeenCalled();
-			expect(declSpy).not.toHaveBeenCalled();
-
-			expect(execa).toHaveBeenCalledWith(
-				'tsc',
-				['--declaration', '--declarationDir', 'dts', '--declarationMap', '--emitDeclarationOnly'],
-				{
-					cwd: fixturePath.path(),
-					preferLocal: true,
-				},
-			);
-		});
-
-		it('without refs, supports a custom config name', async () => {
-			await artifact.build({ declarationConfig: 'tsconfig.build.json' });
-
-			expect(declRootSpy).toHaveBeenCalled();
-			expect(declSpy).not.toHaveBeenCalled();
-
-			expect(execa).toHaveBeenCalledWith(
-				'tsc',
-				[
-					'--declaration',
-					'--declarationDir',
-					'dts',
-					'--declarationMap',
-					'--emitDeclarationOnly',
-					'--project',
-					'tsconfig.build.json',
-				],
-				{
-					cwd: fixturePath.path(),
-					preferLocal: true,
-				},
-			);
-		});
-
-		it('with refs, runs package `tsc` with --build', async () => {
-			tsconfigSpy.mockImplementation(() => ({ projectReferences: [{ path: '..' }] } as any));
-
-			await artifact.build({});
-
-			expect(declRootSpy).not.toHaveBeenCalled();
-			expect(declSpy).toHaveBeenCalled();
-
-			expect(execa).toHaveBeenCalledWith('tsc', ['--build', '--force'], {
-				cwd: fixturePath.path(),
-				preferLocal: true,
-			});
-		});
-
-		it('with refs, supports a custom config name', async () => {
-			tsconfigSpy.mockImplementation(() => ({ projectReferences: [{ path: '..' }] } as any));
-
-			await artifact.build({ declarationConfig: 'tsconfig.build.json' });
-
-			expect(declRootSpy).not.toHaveBeenCalled();
-			expect(declSpy).toHaveBeenCalled();
-
-			expect(execa).toHaveBeenCalledWith('tsc', ['--build', '--force', 'tsconfig.build.json'], {
-				cwd: fixturePath.path(),
-				preferLocal: true,
-			});
-		});
-
-		it('when composite, runs package `tsc` with --build', async () => {
-			tsconfigSpy.mockImplementation(() => ({ options: { composite: true } } as any));
-
-			await artifact.build({});
-
-			expect(declRootSpy).not.toHaveBeenCalled();
-			expect(declSpy).toHaveBeenCalled();
-
-			expect(execa).toHaveBeenCalledWith('tsc', ['--build', '--force'], {
-				cwd: fixturePath.path(),
-				preferLocal: true,
-			});
-		});
-
-		it('when composite, supports a custom config name', async () => {
-			tsconfigSpy.mockImplementation(() => ({ options: { composite: true } } as any));
-
-			await artifact.build({ declarationConfig: 'tsconfig.build.json' });
-
-			expect(declRootSpy).not.toHaveBeenCalled();
-			expect(declSpy).toHaveBeenCalled();
-
-			expect(execa).toHaveBeenCalledWith('tsc', ['--build', '--force', 'tsconfig.build.json'], {
-				cwd: fixturePath.path(),
-				preferLocal: true,
-			});
-		});
-	});
-
 	describe('findEntryPoint()', () => {
 		it('returns mirrored source path', () => {
 			expect(artifact.findEntryPoint('index')).toBe('./dts/index.d.ts');
