--- packages/packemon/tests-old/tsconfig.json
@@ -1,20 +1,0 @@
-{
-  "compilerOptions": {
-    "composite": false,
-    "emitDeclarationOnly": false,
-    "noEmit": true,
-    "rootDir": "."
-  },
-  "extends": "../../../tsconfig.options.json",
-  "include": [
-    "**/*",
-    "../types/**/*",
-    "../../../types/**/*",
-, "../tests/configs.test.ts"    "../tests/PackageValidator.test.ts"
-  ],
-  "references": [
-    {
-      "path": ".."
-    }
-  ]
-}
